Tested on Macbook Pro M1.

Previous (~80ms / token).

```
🔶 Pred   78 ms Sync    0 ms | Sent     0 kB Recv     0 kB |  Tensor
🔶 Pred   87 ms Sync    0 ms | Sent     0 kB Recv     0 kB |  parallel
🔶 Pred   83 ms Sync    0 ms | Sent     0 kB Recv     0 kB | ism
🔶 Pred   88 ms Sync    0 ms | Sent     0 kB Recv     0 kB | ,
🔶 Pred   93 ms Sync    0 ms | Sent     0 kB Recv     0 kB |  a
🔶 Pred   94 ms Sync    0 ms | Sent     0 kB Recv     0 kB |  technique
```

Now (~66ms / token):

```
🔶 Pred   67 ms Sync    0 ms | Sent     0 kB Recv     0 kB | This
🔶 Pred   66 ms Sync    0 ms | Sent     0 kB Recv     0 kB |  note
🔶 Pred   66 ms Sync    0 ms | Sent     0 kB Recv     0 kB |  is
🔶 Pred   66 ms Sync    0 ms | Sent     0 kB Recv     0 kB |  a
🔶 Pred   66 ms Sync    0 ms | Sent     0 kB Recv     0 kB |  rewritten
🔶 Pred   66 ms Sync    0 ms | Sent     0 kB Recv     0 kB |  version
```